[Refactor] 반복문으로 요일cell 추가하도록 수정, DateManager, DummyDataManager 분리
